Programa funcional con requeriemientos basicos
Se tienen los modulos de:

Seguridad
Suministros
Analisis de precios Unitarios
Presupuestos

Todos con su respectivo CRUD

+++ presupuesto/src/main/java/org/jhecohe/repositorio/RoleRepositorio.java
@@ -3,7 +3,7 @@ package org.jhecohe.repositorio;
 import org.jhecohe.dominio.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepositorio extends JpaRepository<Role, Long>{
+public interface RoleRepositorio extends JpaRepository<Role, Integer>{
 
 	Role findByDescripcion(String descripcion);
 }
